Rename deferred publish SDK sources to deferred_publication(s).
Align filenames with the publication handle naming and update includes/ya.make.
